Prepare for trustee generalization, made by the engine
Trustee generalization will require information about the partition
the trustee is in (e.g. to strip its DN into a relative DN within
that partition). We may also want to do this computation in the
engine to be able to generalize when dumping two domains at once
(e.g. see that one delegation is present in the two domains, and
find the most common denominator in them).