Fix ResultLocation serializer, and move to dedicated files
Json serialization doesn't do dynamic type check, it just calls the
serializer it knows for the generic class it's given. So we'll need
to do a switch-case on the dynamic type inside the serializer.

Because of this, the class has grown too much and needs a dedicated
file. Also move ResultTrustee for good measure.