Don't serialize warnings/errors when they are empty
This is just for human-friendlyness purposes, not strictly
required. The JSON serializer is configured to ignore default
values when serializing (this e.g. removes all object ACEs
with their 2 object GUIDs set to Guid.Empty), but an empty list
is an instance, not a default value, it needs to be set to null
in order to be ignored by serialization.

Another approach would be to [JsonIgnore] these properties
themselves, and to create another property with a random name
and a [JsonPropertyName("warnings")].
See https://stackoverflow.com/a/71409972
However the C# serializer seems to be bugged and to use the random
property name when matching constructor arguments to property names...